Continued with the refactoring.
